Change casts to C++ style casts, and delete unnecessary casts.

+++ mert/FeatureStats.cpp
@@ -46,7 +46,7 @@ void SparseVector::set(const string& name, FeatureStatsType value) {
 
 void SparseVector::write(ostream& out, const string& sep) const {
   for (fvector_t::const_iterator i = fvector_.begin(); i != fvector_.end(); ++i) {
-    if (abs((float)(i->second)) < 0.00001) continue;
+    if (abs(i->second) < 0.00001) continue;
     string name = id2name_[i->first];
     out << name << sep << i->second << " ";
   }
